moving to loading file, get default settings in role_model method. set default in model to arctic.Role
