+++ arctic/users/views.py
@@ -2,8 +2,8 @@ from django.contrib.auth import get_user_model
 from django.core.urlresolvers import reverse_lazy
 from django.utils.translation import ugettext as _
 
-from arctic import get_user_role_model
 from arctic.generics import (CreateView, ListView, UpdateView)
+from arctic.loading import get_user_role_model
 
 from .forms import (UserChangeMultiForm, UserCreationMultiForm)
